:building_construction: Refactor hook code to module
